target wasm32 communication module
