+++ binary_port_access/Cargo.toml
@@ -22,7 +22,13 @@ rand = { version = "0.8.5", default-features = false }
 wasm-bindgen = "*"
 wasm-bindgen-futures = "*"
 js-sys = "*"
-web-sys = { version = "0.3", features = ["WebSocket", "MessageEvent"] }
+web-sys = { version = "0.3", features = [
+  "Blob",
+  "FileReader",
+  "ProgressEvent",
+  "WebSocket",
+  "MessageEvent",
+] }
 gloo-utils = { version = "0.2", default-features = false, features = ["serde"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
